Return to storing objects in queue
Thought it would save memory to store only
the tiny coords in the queue and re-build
the combo before returning it, but on further
consideration, that means the combo itself
is built and GC'd twice. The only benefit
to tiny things in the queue is that the
json dump is smaller, but I can just
modify dump/load to do that anyway.